backend/requirements.txt
flask
flask-cors
python-dotenv
groq
requests
reportlab
PyPDF2
gunicorn
